Move tests loadbalancing
