+++ tests/tests_network/mod.rs
@@ -1,0 +1,2 @@
+mod canonical_url;
+mod loadbalancing;
